+++ example/lib/model/chinook.g.dart
@@ -730,7 +730,8 @@ class Album extends TableBase {
 
   /// saveAll method saves the sent List<Album> as a bulk in one transaction
   /// Returns a <List<BoolResult>>
-  static Future<List<dynamic>> saveAll(List<Album> albums) async {
+  static Future<List<dynamic>> saveAll(List<Album> albums,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     List<dynamic>? result = [];
     // If there is no open transaction, start one
     final isStartedBatch = await Chinookdb().batchStart();
@@ -738,7 +739,10 @@ class Album extends TableBase {
       await obj.save(ignoreBatch: false);
     }
     if (!isStartedBatch) {
-      result = await Chinookdb().batchCommit();
+      result = await Chinookdb().batchCommit(
+          exclusive: exclusive,
+          noResult: noResult,
+          continueOnError: continueOnError);
       for (int i = 0; i < albums.length; i++) {
         if (albums[i].AlbumId == null) {
           albums[i].AlbumId = result![i] as int;
@@ -779,10 +783,14 @@ class Album extends TableBase {
   /// upsertAll() method is faster then saveAll() method. upsertAll() should be used when you are sure that the primary key is greater than zero
   /// Returns a BoolCommitResult
   @override
-  Future<BoolCommitResult> upsertAll(List<Album> albums) async {
+  Future<BoolCommitResult> upsertAll(List<Album> albums,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     final results = await _mnAlbum.rawInsertAll(
         'INSERT OR REPLACE INTO Album (AlbumId, Title, ArtistId)  VALUES (?,?,?)',
-        albums);
+        albums,
+        exclusive: exclusive,
+        noResult: noResult,
+        continueOnError: continueOnError);
     return results;
   }
 
@@ -1248,7 +1256,7 @@ class AlbumFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<Album>> items = []..add(DropdownMenuItem(
         value: Album(),
-        child: Text('Select Album'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -1277,7 +1285,7 @@ class AlbumFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<int>> items = []..add(DropdownMenuItem(
         value: 0,
-        child: Text('Select Album'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -1659,7 +1667,8 @@ class Artist extends TableBase {
 
   /// saveAll method saves the sent List<Artist> as a bulk in one transaction
   /// Returns a <List<BoolResult>>
-  static Future<List<dynamic>> saveAll(List<Artist> artists) async {
+  static Future<List<dynamic>> saveAll(List<Artist> artists,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     List<dynamic>? result = [];
     // If there is no open transaction, start one
     final isStartedBatch = await Chinookdb().batchStart();
@@ -1667,7 +1676,10 @@ class Artist extends TableBase {
       await obj.save(ignoreBatch: false);
     }
     if (!isStartedBatch) {
-      result = await Chinookdb().batchCommit();
+      result = await Chinookdb().batchCommit(
+          exclusive: exclusive,
+          noResult: noResult,
+          continueOnError: continueOnError);
       for (int i = 0; i < artists.length; i++) {
         if (artists[i].ArtistId == null) {
           artists[i].ArtistId = result![i] as int;
@@ -1708,10 +1720,13 @@ class Artist extends TableBase {
   /// upsertAll() method is faster then saveAll() method. upsertAll() should be used when you are sure that the primary key is greater than zero
   /// Returns a BoolCommitResult
   @override
-  Future<BoolCommitResult> upsertAll(List<Artist> artists) async {
+  Future<BoolCommitResult> upsertAll(List<Artist> artists,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     final results = await _mnArtist.rawInsertAll(
-        'INSERT OR REPLACE INTO Artist (ArtistId, Name)  VALUES (?,?)',
-        artists);
+        'INSERT OR REPLACE INTO Artist (ArtistId, Name)  VALUES (?,?)', artists,
+        exclusive: exclusive,
+        noResult: noResult,
+        continueOnError: continueOnError);
     return results;
   }
 
@@ -2163,7 +2178,7 @@ class ArtistFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<Artist>> items = []..add(DropdownMenuItem(
         value: Artist(),
-        child: Text('Select Artist'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -2192,7 +2207,7 @@ class ArtistFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<int>> items = []..add(DropdownMenuItem(
         value: 0,
-        child: Text('Select Artist'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -2795,7 +2810,8 @@ class Customer extends TableBase {
 
   /// saveAll method saves the sent List<Customer> as a bulk in one transaction
   /// Returns a <List<BoolResult>>
-  static Future<List<dynamic>> saveAll(List<Customer> customers) async {
+  static Future<List<dynamic>> saveAll(List<Customer> customers,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     List<dynamic>? result = [];
     // If there is no open transaction, start one
     final isStartedBatch = await Chinookdb().batchStart();
@@ -2803,7 +2819,10 @@ class Customer extends TableBase {
       await obj.save(ignoreBatch: false);
     }
     if (!isStartedBatch) {
-      result = await Chinookdb().batchCommit();
+      result = await Chinookdb().batchCommit(
+          exclusive: exclusive,
+          noResult: noResult,
+          continueOnError: continueOnError);
       for (int i = 0; i < customers.length; i++) {
         if (customers[i].CustomerId == null) {
           customers[i].CustomerId = result![i] as int;
@@ -2859,10 +2878,14 @@ class Customer extends TableBase {
   /// upsertAll() method is faster then saveAll() method. upsertAll() should be used when you are sure that the primary key is greater than zero
   /// Returns a BoolCommitResult
   @override
-  Future<BoolCommitResult> upsertAll(List<Customer> customers) async {
+  Future<BoolCommitResult> upsertAll(List<Customer> customers,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     final results = await _mnCustomer.rawInsertAll(
         'INSERT OR REPLACE INTO Customer (CustomerId, FirstName, LastName, Company, Address, City, State, Country, PostalCode, Phone, Fax, Email, SupportRepId)  VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?)',
-        customers);
+        customers,
+        exclusive: exclusive,
+        noResult: noResult,
+        continueOnError: continueOnError);
     return results;
   }
 
@@ -3383,7 +3406,7 @@ class CustomerFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<Customer>> items = []..add(DropdownMenuItem(
         value: Customer(),
-        child: Text('Select Customer'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -3412,7 +3435,7 @@ class CustomerFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<int>> items = []..add(DropdownMenuItem(
         value: 0,
-        child: Text('Select Customer'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -4180,7 +4203,8 @@ class Employee extends TableBase {
 
   /// saveAll method saves the sent List<Employee> as a bulk in one transaction
   /// Returns a <List<BoolResult>>
-  static Future<List<dynamic>> saveAll(List<Employee> employees) async {
+  static Future<List<dynamic>> saveAll(List<Employee> employees,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     List<dynamic>? result = [];
     // If there is no open transaction, start one
     final isStartedBatch = await Chinookdb().batchStart();
@@ -4188,7 +4212,10 @@ class Employee extends TableBase {
       await obj.save(ignoreBatch: false);
     }
     if (!isStartedBatch) {
-      result = await Chinookdb().batchCommit();
+      result = await Chinookdb().batchCommit(
+          exclusive: exclusive,
+          noResult: noResult,
+          continueOnError: continueOnError);
       for (int i = 0; i < employees.length; i++) {
         if (employees[i].EmployeeId == null) {
           employees[i].EmployeeId = result![i] as int;
@@ -4246,10 +4273,14 @@ class Employee extends TableBase {
   /// upsertAll() method is faster then saveAll() method. upsertAll() should be used when you are sure that the primary key is greater than zero
   /// Returns a BoolCommitResult
   @override
-  Future<BoolCommitResult> upsertAll(List<Employee> employees) async {
+  Future<BoolCommitResult> upsertAll(List<Employee> employees,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     final results = await _mnEmployee.rawInsertAll(
         'INSERT OR REPLACE INTO Employee (EmployeeId, LastName, FirstName, Title, BirthDate, HireDate, Address, City, State, Country, PostalCode, Phone, Fax, Email, ReportsTo)  VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)',
-        employees);
+        employees,
+        exclusive: exclusive,
+        noResult: noResult,
+        continueOnError: continueOnError);
     return results;
   }
 
@@ -4815,7 +4846,7 @@ class EmployeeFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<Employee>> items = []..add(DropdownMenuItem(
         value: Employee(),
-        child: Text('Select Employee'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -4844,7 +4875,7 @@ class EmployeeFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<int>> items = []..add(DropdownMenuItem(
         value: 0,
-        child: Text('Select Employee'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -5297,7 +5328,8 @@ class Genre extends TableBase {
 
   /// saveAll method saves the sent List<Genre> as a bulk in one transaction
   /// Returns a <List<BoolResult>>
-  static Future<List<dynamic>> saveAll(List<Genre> genres) async {
+  static Future<List<dynamic>> saveAll(List<Genre> genres,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     List<dynamic>? result = [];
     // If there is no open transaction, start one
     final isStartedBatch = await Chinookdb().batchStart();
@@ -5305,7 +5337,10 @@ class Genre extends TableBase {
       await obj.save(ignoreBatch: false);
     }
     if (!isStartedBatch) {
-      result = await Chinookdb().batchCommit();
+      result = await Chinookdb().batchCommit(
+          exclusive: exclusive,
+          noResult: noResult,
+          continueOnError: continueOnError);
       for (int i = 0; i < genres.length; i++) {
         if (genres[i].GenreId == null) {
           genres[i].GenreId = result![i] as int;
@@ -5346,9 +5381,13 @@ class Genre extends TableBase {
   /// upsertAll() method is faster then saveAll() method. upsertAll() should be used when you are sure that the primary key is greater than zero
   /// Returns a BoolCommitResult
   @override
-  Future<BoolCommitResult> upsertAll(List<Genre> genres) async {
+  Future<BoolCommitResult> upsertAll(List<Genre> genres,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     final results = await _mnGenre.rawInsertAll(
-        'INSERT OR REPLACE INTO Genre (GenreId, Name)  VALUES (?,?)', genres);
+        'INSERT OR REPLACE INTO Genre (GenreId, Name)  VALUES (?,?)', genres,
+        exclusive: exclusive,
+        noResult: noResult,
+        continueOnError: continueOnError);
     return results;
   }
 
@@ -5798,7 +5837,7 @@ class GenreFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<Genre>> items = []..add(DropdownMenuItem(
         value: Genre(),
-        child: Text('Select Genre'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -5827,7 +5866,7 @@ class GenreFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<int>> items = []..add(DropdownMenuItem(
         value: 0,
-        child: Text('Select Genre'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -6386,7 +6425,8 @@ class Invoice extends TableBase {
 
   /// saveAll method saves the sent List<Invoice> as a bulk in one transaction
   /// Returns a <List<BoolResult>>
-  static Future<List<dynamic>> saveAll(List<Invoice> invoices) async {
+  static Future<List<dynamic>> saveAll(List<Invoice> invoices,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     List<dynamic>? result = [];
     // If there is no open transaction, start one
     final isStartedBatch = await Chinookdb().batchStart();
@@ -6394,7 +6434,10 @@ class Invoice extends TableBase {
       await obj.save(ignoreBatch: false);
     }
     if (!isStartedBatch) {
-      result = await Chinookdb().batchCommit();
+      result = await Chinookdb().batchCommit(
+          exclusive: exclusive,
+          noResult: noResult,
+          continueOnError: continueOnError);
       for (int i = 0; i < invoices.length; i++) {
         if (invoices[i].InvoiceId == null) {
           invoices[i].InvoiceId = result![i] as int;
@@ -6446,10 +6489,14 @@ class Invoice extends TableBase {
   /// upsertAll() method is faster then saveAll() method. upsertAll() should be used when you are sure that the primary key is greater than zero
   /// Returns a BoolCommitResult
   @override
-  Future<BoolCommitResult> upsertAll(List<Invoice> invoices) async {
+  Future<BoolCommitResult> upsertAll(List<Invoice> invoices,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     final results = await _mnInvoice.rawInsertAll(
         'INSERT OR REPLACE INTO Invoice (InvoiceId, InvoiceDate, BillingAddress, BillingCity, BillingState, BillingCountry, BillingPostalCode, Total, CustomerId)  VALUES (?,?,?,?,?,?,?,?,?)',
-        invoices);
+        invoices,
+        exclusive: exclusive,
+        noResult: noResult,
+        continueOnError: continueOnError);
     return results;
   }
 
@@ -6954,7 +7001,7 @@ class InvoiceFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<Invoice>> items = []..add(DropdownMenuItem(
         value: Invoice(),
-        child: Text('Select Invoice'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -6983,7 +7030,7 @@ class InvoiceFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<int>> items = []..add(DropdownMenuItem(
         value: 0,
-        child: Text('Select Invoice'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -7475,7 +7522,8 @@ class InvoiceLine extends TableBase {
 
   /// saveAll method saves the sent List<InvoiceLine> as a bulk in one transaction
   /// Returns a <List<BoolResult>>
-  static Future<List<dynamic>> saveAll(List<InvoiceLine> invoicelines) async {
+  static Future<List<dynamic>> saveAll(List<InvoiceLine> invoicelines,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     List<dynamic>? result = [];
     // If there is no open transaction, start one
     final isStartedBatch = await Chinookdb().batchStart();
@@ -7483,7 +7531,10 @@ class InvoiceLine extends TableBase {
       await obj.save(ignoreBatch: false);
     }
     if (!isStartedBatch) {
-      result = await Chinookdb().batchCommit();
+      result = await Chinookdb().batchCommit(
+          exclusive: exclusive,
+          noResult: noResult,
+          continueOnError: continueOnError);
       for (int i = 0; i < invoicelines.length; i++) {
         if (invoicelines[i].InvoiceLineId == null) {
           invoicelines[i].InvoiceLineId = result![i] as int;
@@ -7526,10 +7577,14 @@ class InvoiceLine extends TableBase {
   /// upsertAll() method is faster then saveAll() method. upsertAll() should be used when you are sure that the primary key is greater than zero
   /// Returns a BoolCommitResult
   @override
-  Future<BoolCommitResult> upsertAll(List<InvoiceLine> invoicelines) async {
+  Future<BoolCommitResult> upsertAll(List<InvoiceLine> invoicelines,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     final results = await _mnInvoiceLine.rawInsertAll(
         'INSERT OR REPLACE INTO InvoiceLine (InvoiceLineId, UnitPrice, Quantity, TrackId, InvoiceId)  VALUES (?,?,?,?,?)',
-        invoicelines);
+        invoicelines,
+        exclusive: exclusive,
+        noResult: noResult,
+        continueOnError: continueOnError);
     return results;
   }
 
@@ -7985,7 +8040,7 @@ class InvoiceLineFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<InvoiceLine>> items = []..add(DropdownMenuItem(
         value: InvoiceLine(),
-        child: Text('Select InvoiceLine'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -8014,7 +8069,7 @@ class InvoiceLineFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<int>> items = []..add(DropdownMenuItem(
         value: 0,
-        child: Text('Select InvoiceLine'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -8410,7 +8465,8 @@ class MediaType extends TableBase {
 
   /// saveAll method saves the sent List<MediaType> as a bulk in one transaction
   /// Returns a <List<BoolResult>>
-  static Future<List<dynamic>> saveAll(List<MediaType> mediatypes) async {
+  static Future<List<dynamic>> saveAll(List<MediaType> mediatypes,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     List<dynamic>? result = [];
     // If there is no open transaction, start one
     final isStartedBatch = await Chinookdb().batchStart();
@@ -8418,7 +8474,10 @@ class MediaType extends TableBase {
       await obj.save(ignoreBatch: false);
     }
     if (!isStartedBatch) {
-      result = await Chinookdb().batchCommit();
+      result = await Chinookdb().batchCommit(
+          exclusive: exclusive,
+          noResult: noResult,
+          continueOnError: continueOnError);
       for (int i = 0; i < mediatypes.length; i++) {
         if (mediatypes[i].MediaTypeId == null) {
           mediatypes[i].MediaTypeId = result![i] as int;
@@ -8460,10 +8519,14 @@ class MediaType extends TableBase {
   /// upsertAll() method is faster then saveAll() method. upsertAll() should be used when you are sure that the primary key is greater than zero
   /// Returns a BoolCommitResult
   @override
-  Future<BoolCommitResult> upsertAll(List<MediaType> mediatypes) async {
+  Future<BoolCommitResult> upsertAll(List<MediaType> mediatypes,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     final results = await _mnMediaType.rawInsertAll(
         'INSERT OR REPLACE INTO MediaType (MediaTypeId, Name)  VALUES (?,?)',
-        mediatypes);
+        mediatypes,
+        exclusive: exclusive,
+        noResult: noResult,
+        continueOnError: continueOnError);
     return results;
   }
 
@@ -8921,7 +8984,7 @@ class MediaTypeFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<MediaType>> items = []..add(DropdownMenuItem(
         value: MediaType(),
-        child: Text('Select MediaType'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -8950,7 +9013,7 @@ class MediaTypeFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<int>> items = []..add(DropdownMenuItem(
         value: 0,
-        child: Text('Select MediaType'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -9324,7 +9387,8 @@ class Playlist extends TableBase {
 
   /// saveAll method saves the sent List<Playlist> as a bulk in one transaction
   /// Returns a <List<BoolResult>>
-  static Future<List<dynamic>> saveAll(List<Playlist> playlists) async {
+  static Future<List<dynamic>> saveAll(List<Playlist> playlists,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     List<dynamic>? result = [];
     // If there is no open transaction, start one
     final isStartedBatch = await Chinookdb().batchStart();
@@ -9332,7 +9396,10 @@ class Playlist extends TableBase {
       await obj.save(ignoreBatch: false);
     }
     if (!isStartedBatch) {
-      result = await Chinookdb().batchCommit();
+      result = await Chinookdb().batchCommit(
+          exclusive: exclusive,
+          noResult: noResult,
+          continueOnError: continueOnError);
       for (int i = 0; i < playlists.length; i++) {
         if (playlists[i].PlaylistId == null) {
           playlists[i].PlaylistId = result![i] as int;
@@ -9374,10 +9441,14 @@ class Playlist extends TableBase {
   /// upsertAll() method is faster then saveAll() method. upsertAll() should be used when you are sure that the primary key is greater than zero
   /// Returns a BoolCommitResult
   @override
-  Future<BoolCommitResult> upsertAll(List<Playlist> playlists) async {
+  Future<BoolCommitResult> upsertAll(List<Playlist> playlists,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     final results = await _mnPlaylist.rawInsertAll(
         'INSERT OR REPLACE INTO Playlist (PlaylistId, Name)  VALUES (?,?)',
-        playlists);
+        playlists,
+        exclusive: exclusive,
+        noResult: noResult,
+        continueOnError: continueOnError);
     return results;
   }
 
@@ -9810,7 +9881,7 @@ class PlaylistFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<Playlist>> items = []..add(DropdownMenuItem(
         value: Playlist(),
-        child: Text('Select Playlist'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -9839,7 +9910,7 @@ class PlaylistFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<int>> items = []..add(DropdownMenuItem(
         value: 0,
-        child: Text('Select Playlist'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -10529,7 +10600,8 @@ class Track extends TableBase {
 
   /// saveAll method saves the sent List<Track> as a bulk in one transaction
   /// Returns a <List<BoolResult>>
-  static Future<List<dynamic>> saveAll(List<Track> tracks) async {
+  static Future<List<dynamic>> saveAll(List<Track> tracks,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     List<dynamic>? result = [];
     // If there is no open transaction, start one
     final isStartedBatch = await Chinookdb().batchStart();
@@ -10537,7 +10609,10 @@ class Track extends TableBase {
       await obj.save(ignoreBatch: false);
     }
     if (!isStartedBatch) {
-      result = await Chinookdb().batchCommit();
+      result = await Chinookdb().batchCommit(
+          exclusive: exclusive,
+          noResult: noResult,
+          continueOnError: continueOnError);
       for (int i = 0; i < tracks.length; i++) {
         if (tracks[i].TrackId == null) {
           tracks[i].TrackId = result![i] as int;
@@ -10588,10 +10663,14 @@ class Track extends TableBase {
   /// upsertAll() method is faster then saveAll() method. upsertAll() should be used when you are sure that the primary key is greater than zero
   /// Returns a BoolCommitResult
   @override
-  Future<BoolCommitResult> upsertAll(List<Track> tracks) async {
+  Future<BoolCommitResult> upsertAll(List<Track> tracks,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     final results = await _mnTrack.rawInsertAll(
         'INSERT OR REPLACE INTO Track (TrackId, Name, Composer, Milliseconds, Bytes, UnitPrice, MediaTypeId, GenreId, AlbumId)  VALUES (?,?,?,?,?,?,?,?,?)',
-        tracks);
+        tracks,
+        exclusive: exclusive,
+        noResult: noResult,
+        continueOnError: continueOnError);
     return results;
   }
 
@@ -11142,7 +11221,7 @@ class TrackFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<Track>> items = []..add(DropdownMenuItem(
         value: Track(),
-        child: Text('Select Track'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -11171,7 +11250,7 @@ class TrackFilterBuilder extends ConjunctionBase {
     final int count = data.length;
     final List<DropdownMenuItem<int>> items = []..add(DropdownMenuItem(
         value: 0,
-        child: Text('Select Track'),
+        child: Text('-'),
       ));
     for (int i = 0; i < count; i++) {
       items.add(
@@ -12279,7 +12358,7 @@ class PlaylistTrack extends TableBase {
     final result = BoolResult(success: false);
     try {
       await _mnPlaylistTrack.rawInsert(
-          'INSERT ${isSaved! ? 'OR REPLACE' : ''} INTO PlaylistTrack ( TrackId, PlaylistId)  VALUES ()',
+          'INSERT ${isSaved! ? 'OR REPLACE' : ''} INTO PlaylistTrack ( TrackId, PlaylistId)  VALUES (?,?)',
           toArgsWithIds(),
           ignoreBatch);
       result.success = true;
@@ -12294,8 +12373,8 @@ class PlaylistTrack extends TableBase {
 
   /// saveAll method saves the sent List<PlaylistTrack> as a bulk in one transaction
   /// Returns a <List<BoolResult>>
-  static Future<List<dynamic>> saveAll(
-      List<PlaylistTrack> playlisttracks) async {
+  static Future<List<dynamic>> saveAll(List<PlaylistTrack> playlisttracks,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     List<dynamic>? result = [];
     // If there is no open transaction, start one
     final isStartedBatch = await Chinookdb().batchStart();
@@ -12303,7 +12382,10 @@ class PlaylistTrack extends TableBase {
       await obj.save();
     }
     if (!isStartedBatch) {
-      result = await Chinookdb().batchCommit();
+      result = await Chinookdb().batchCommit(
+          exclusive: exclusive,
+          noResult: noResult,
+          continueOnError: continueOnError);
     }
     return result!;
   }
@@ -12314,7 +12396,7 @@ class PlaylistTrack extends TableBase {
   Future<int?> upsert({bool ignoreBatch = true}) async {
     try {
       final result = await _mnPlaylistTrack.rawInsert(
-          'INSERT OR REPLACE INTO PlaylistTrack ( TrackId, PlaylistId)  VALUES ()',
+          'INSERT OR REPLACE INTO PlaylistTrack ( TrackId, PlaylistId)  VALUES (?,?)',
           [TrackId, PlaylistId],
           ignoreBatch);
       if (result! > 0) {
@@ -12340,10 +12422,14 @@ class PlaylistTrack extends TableBase {
   /// upsertAll() method is faster then saveAll() method. upsertAll() should be used when you are sure that the primary key is greater than zero
   /// Returns a BoolCommitResult
   @override
-  Future<BoolCommitResult> upsertAll(List<PlaylistTrack> playlisttracks) async {
+  Future<BoolCommitResult> upsertAll(List<PlaylistTrack> playlisttracks,
+      {bool? exclusive, bool? noResult, bool? continueOnError}) async {
     final results = await _mnPlaylistTrack.rawInsertAll(
-        'INSERT OR REPLACE INTO PlaylistTrack ( TrackId, PlaylistId)  VALUES ()',
-        playlisttracks);
+        'INSERT OR REPLACE INTO PlaylistTrack ( TrackId, PlaylistId)  VALUES (?,?)',
+        playlisttracks,
+        exclusive: exclusive,
+        noResult: noResult,
+        continueOnError: continueOnError);
     return results;
   }