+++ sqfentity_gen/CHANGELOG.md
@@ -1,4 +1,31 @@
-## 2.2.0+7
+## 2.3.0+4
+- fixed issue [281](https://github.com/hhtokpinar/sqfEntity/issues/281)
+  and removed customizing datetime formats.
+  If you want to use DefaultDateTime format. Specify the sqfentity_gen version in pubspec.yaml
+
+  sqfentity_gen: 2.3.0+3
+
+
+## 2.3.0+3
+- fixed issue [279](https://github.com/hhtokpinar/sqfEntity/issues/279)
+
+## 2.3.0+1
+- fixed issue [278](https://github.com/hhtokpinar/sqfEntity/issues/278)
+
+## 2.3.0
+- New global default Date/Time Formats
+- Making the viewAdd takes optional object
+- some minor updates
+
+## 2.2.0+12
+- Added optional parameters into upsertAll method for batchCommit 
+
+## 2.2.0+11
+- Added optional parameters into saveAll method for batchCommit 
+
+   saveAll({bool? exclusive, bool? noResult, bool? continueOnError})
+
+## 2.2.0+8
 - Removed dependency of flutter_datetime_picker
 
 created UITools.showDateTimePicker() method to support another package using instead